+++ CHANGELOG.rst
@@ -34,4 +34,5 @@ first official Alpha release.
  * Introduced the pagecache object for caching Pages
  * Sitemap handled by pagecache. For now only static pages and application pages without arguments are included.
  * Added lloogg_analytics and css_validators in template tags
- * NEW FIELD IN PAGES!! Added doctype field for specifying document type (HTML 4.01, XHTML 1, HTML 5)
+ * NEW FIELD IN PAGE MODEL!! Added doctype field for specifying document type (HTML 4.01, XHTML 1, HTML 5)
+ * NEW FIELD IN PAGE MODEL!! Added insidemap for disabling a page from sidemap and robots